requirements.txt
faiss-cpu
numpy
sentence-transformers
streamlit
unidecode